Allow rubocop 1.0 to be used
 - My project all use the most recent version of rubocop.  The current version of solargraph (0.39.17) does not work against rubocop beyond version 1. This means that although it seems solargraph start correctly the solargraph process will just hang on "init" and does not warn/autoformat.

Update the gemspec file and create a new gem with this command `gem build solargraph.gemspec`. This results in the solargraph-0.40.0.gem which can be installed directly. For some project it's easiest to copy the .gem file into the working directory of the project and run `gem install solargraph`. You should see this version being installed. After that it's fine to remove the solargraph gem from this direcoty.